Decouple Capture Context tests
The "capture code context spec" currently depends on code search which can change as the algorithm evolves. This makes development hard since it's unclear when those tests fail if it's because the functionality changed or the search output changed. Also since the existing "capture context" logic is based on quirks of how the code search algorithm works, when the search is improved it may need different "capture context" logic before it produces good output.

For these cases, we actually need to duplicate the tests 3 times.

- Integration with everything
- Capture Code Context works with static input
- CodeSearch produces the static input we expect

So when all 3 start failing then we can look at it in isolation and see if it "makes sense". I.e. if the CodeSearch output actually "improved" but the integration and capture context are failing then we need to write new logic.