Initial idea to search by "diagnosing" a node
I am very happy with this direction.

When looking at a node, we can determine if it holds a syntax error or not. If it's invalid it means that either one of its parents are the problem or it is.

Through test based exploration I found several properties that seem to indicate a specific case happened. For instance if only one parent is invalid in isolation, then we should explore just that parent.

I wanted to decouple inspection from modification so that we can look at a node to guess what's wrong with it, without having to take the next step of actually taking an action
to return the invalid inner/parent node. This produces some possible performance problems as some of the checks for diagnose may not be cheap. I'm mitigating that where possible via memoizing expensive values. It's not ideal, but it's holding up okay.

This strategy handles all of our problems like a champ, except for the full `|` example which needs more investigation.

Note to future self we also need to handle:

- actually investigating multiple syntax errors, like if there are two syntax errors in a document
- Handling the `else` case. This is where kw/end are matched but there's one or more dangling pseudo keywords ("ensure", "rescue", "else", "elsif", "when", "retry") left in the middle.