Implement tracing on new search
Previously I implemented recording/tracing on building the tree, we also need to introspect how the tree is searched. This commit adds that tracing and DRYs up the code a bit.

I don't love that there's this DEFAULT_VALUE that needs to be understood by so many abstraction layers. However if we want a false value like `nil` to represent "default state" then it seems to be needed.

We also want to be as lazy as possible about tracing because each individual level/layer might need to be invoked with the debug env vars. This is good enough for now.

The only other major downside is that the performance of this is quite expensive. For a search that takes half a second with tracing off it takes 1 min 47 seconds with it on.